Buffer overflow
Example_segmentationFault2.c is non compiled file. It is a program that causes a segmentation fault. There's a buffer buf with 3 positions and it is asked to enter a postal code that will be stored in buf. Run it after compiling. When a postal code of 5 numbers is entered it will overwrite a part of memory that is restricted and will show error: ***stack smashing detected***